refactor(api): extract FaceDetectionCore shared by photo/video face detection
Moves bbox normalization, L2 embedding normalization, Rekognition throttle
mapping, Face row creation + person-matching loop, and MediaFaceStatus helpers
(markProcessing / markStatus / markFailed) out of FaceDetectionService and into
a new injectable FaceDetectionCore service.

FaceDetectionService is refactored to delegate to the core; all externally
visible behaviour (job processing, status lifecycle, error propagation) is
identical to before.

Notes: integration tests not run — no DB in this environment; `npx tsc --noEmit`
passes clean.
